Module boomphf::hashmap

HashMap data structures, using MPHFs to encode the position of each key in a dense array.

Structs

Boom2Iterator

BoomHashMap
A HashMap data structure where the mapping between keys and values is encoded in a Mphf. This lets us store the keys and values in dense arrays, with ~3 bits/item overhead in the Mphf.

BoomHashMap2
A HashMap data structure where the mapping between keys and 2 values is encoded in a Mphf. You should usually use BoomHashMap with a tuple/struct value type. If the layout overhead of the struct / tuple must be avoided, this variant of is an alternative. This lets us store the keys and values in dense arrays, with ~3 bits/item overhead in the Mphf.

BoomIterator
Iterate over key-value pairs in a BoomHashMap

NoKeyBoomHashMap
A HashMap data structure where the mapping between keys and values is encoded in a Mphf. Keys are not stored - this can greatly improve the memory consumption, but can only be used if you can guarantee that you will only query for keys that were in the original set. Querying for a new key will return a random value, silently.

NoKeyBoomHashMap2
A HashMap data structure where the mapping between keys and values is encoded in a Mphf. Keys are not stored - this can greatly improve the memory consumption, but can only be used if you can guarantee that you will only query for keys that were in the original set. Querying for a new key will return a random value, silently.