Struct boomphf::Mphf

pub struct Mphf<T> { /* fields omitted */ }

A minimal perfect hash function over a set of objects of type T.

Methods

impl<T: Hash + Clone + Debug> Mphf<T>

pub fn new(gamma: f64, objects: &Vec<T>) -> Mphf<T>

Generate a minimal perfect hash function for the set of objects. objects must not contain any duplicate items. gamma controls the tradeoff between the construction-time and run-time speed, and the size of the datastructure representing the hash function. See the paper for details. max_iters - None to never stop trying to find a perfect hash (safe if no duplicates).

pub fn hash(&self, item: &T) -> u64

Compute the hash value of item. This method should only be used with items known to be in construction set. Use try_hash if you cannot guarantee that item was in the construction set. If item was not present in the construction set this function may panic.

pub fn try_hash(&self, item: &T) -> Option<u64>

Compute the hash value of item. If item was not present in the set of objects used to construct the hash function, the return value will an arbitrary value Some(x), or None.

impl<T: Hash + Clone + Debug + Sync + Send> Mphf<T>

pub fn new_parallel(
gamma: f64,
objects: &Vec<T>,
starting_seed: Option<u64>
) -> Mphf<T>

Same as new, but parallelizes work on the rayon default Rayon threadpool. Configure the number of threads on that threadpool to control CPU usage.